correct helpers file name format and move mosendo project to mobile tab
