+++ src/components/SubHeader/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import { Link } from 'react-router-dom'
 
-import ServicesData from 'utils/Services'
+import ServicesData from 'utils/servicesMenuContent'
 import arrowRight from 'assets/images/arrow-right.svg'
 import arrowDown from 'assets/images/chevron-down-solid.svg'
