+++ src/components/Tabs/TechTabs.js
@@ -5,9 +5,9 @@ import TabHeader from './TabHeader'
 import Tabs from './Tabs'
 
 import './index.sass'
-import FrontEndTechs from 'utils/FrontendTechs'
-import BackEndTechs from 'utils/BackEndTechs'
-import MobileStackTechs from 'utils/MobileTechs'
+import FrontEndTechs from 'utils/frontendTechs'
+import BackEndTechs from 'utils/backEndTechs'
+import MobileStackTechs from 'utils/mobileTechs'
 
 class TabContainer extends Component {
   render() {
