+++ src/pages/App/index.js
@@ -5,7 +5,7 @@ import Home from '../Home'
 import About from '../About'
 import TechnologyContainer from '../Technology'
 import Services from '../Services'
-import singleServiceDetails from 'utils/SingleServiceDetails'
+import singleServiceDetails from 'utils/singleServiceDetails'
 import portfolioDetails from 'utils/portfolioDetails'
 import objectToArray from 'utils/objectToArray'
 import ServicesSubPage from '../ServicesSubPage'
