+++ src/pages/Services/components/ServicesDetail/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 
-import { ServicesDetailContent } from 'utils/ServicesDetail'
+import { ServicesDetailContent } from 'utils/servicesDetail'
 import './index.sass'
 import objectToArray from 'utils/objectToArray'
