+++ src/utils/mobileTechs.js
