+++ src/utils/partnersLogo.js
